backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
aiosqlite>=0.19.0
pyjwt>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.6
